comment model moved to folder, page model added and pageService interface added
