+++ server/src/modules/comments/models/index.ts
@@ -1,0 +1,1 @@
+export type {Comment} from "./Comment";
