+++ server/src/modules/pages/index.ts
@@ -1,0 +1,10 @@
+import { Page } from "./models";
+
+export type { Page };
+
+export abstract class PageService {
+  abstract getTopCommentedPages(
+    commentsMinVoteSum: number,
+    limit: number
+  ): Promise<Array<Page>>;
+}
