+++ server/src/modules/pages/models/Page.ts
@@ -1,0 +1,5 @@
+export interface Page {
+  id: string;
+  websiteId: string;
+  commentsCount: number;
+}
