+++ server/src/modules/pages/models/index.ts
@@ -1,0 +1,1 @@
+export type {Page} from "./Page";
